Implementa registo manual de histórico no backend e frontend

+++ frontend/frontend/src/Components/FormTask.jsx
@@ -33,7 +33,7 @@ function FormTask(listMembers) {
   };
 
   const formatDate = (dateString) => {
-    const options = { year: "numeric", month: "long", day: "numeric" };
+    const options = { year: "numeric", month: "numeric", day: "numeric" };
     return new Date(dateString).toLocaleDateString(undefined, options);
   };
